!refactor: remove barrel file exports
`/composables/index.ts` and `/plugins/index.ts` lead to bugs where unused code is pulled into the bundle.

To prevent this from happening, we have decided to remove the barrel exports entirely. This is somewhat of a breaking change, since all composables and plugins no need to be imported directly from their respective folder instead of the barrel file.